feat(generation): restructure web layout column with params overlay menu
- 左栏顺序：尺寸 → 种子 → 提示词 → 反推/图生图/风格迁移/精准参考
- 参数（模型/采样器/噪声调度/步数/CFG/CFG Rescale）改为二级菜单：
  触发条固定在生成控制条上方，面板向上弹出、悬浮于内容之上（带遮罩）
- ParameterPanel 拆分为可组合分节（generation_param_sections），经典布局复用同一批控件
- 修复种子锁定按钮视觉反馈不刷新（经典与官网布局均受影响）
- 清理死代码：showInput 块、selectParameterPanelViewData、inBottomSheet 参数、
  模型抽屉组件与其持久化字段
