+++ lib/core/constants/storage_keys.dart
@@ -51,7 +51,6 @@ class StorageKeys {
   static const String generationLayoutMode = 'generation_layout_mode';
   static const String webLeftPanelWidth = 'web_left_panel_width';
   static const String webLeftPanelExpanded = 'web_left_panel_expanded';
-  static const String webModelDrawerExpanded = 'web_model_drawer_expanded';
   static const String fixedTagsSidebarExpanded = 'fixed_tags_sidebar_expanded';
   static const String fixedTagsSidebarWidth = 'fixed_tags_sidebar_width';
   static const String fixedTagsSidebarViewMode = 'fixed_tags_sidebar_view_mode';
