+++ lib/presentation/screens/generation/widgets/parameter_panel.dart
@@ -1,469 +1,64 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../core/constants/api_constants.dart';
 import '../../../../core/utils/localization_extension.dart';
-import '../../../providers/generation/generation_params_selectors.dart';
 import '../../../providers/image_generation_provider.dart';
-import '../../../providers/krita/krita_bridge_notifier.dart';
-import '../../../providers/layout_state_provider.dart';
-import '../../../utils/asset_protection_guard.dart';
-import '../../../widgets/common/themed_dropdown.dart';
-import '../../../widgets/common/themed_input.dart';
-import '../../../widgets/common/themed_button.dart';
-import '../../../widgets/common/themed_slider.dart';
-import '../../../widgets/common/themed_divider.dart';
-import 'collapsible_model_selector.dart';
-import 'size_selector.dart';
+import 'generation_param_sections.dart';
 import 'img2img_panel.dart';
+import 'precise_reference_panel.dart';
 import 'reverse_prompt_panel.dart';
 import 'unified_reference_panel.dart';
-import 'precise_reference_panel.dart';
-import 'prompt_input.dart';
 
-import '../../../widgets/common/app_toast.dart';
-
-/// 参数面板组件
-class ParameterPanel extends ConsumerStatefulWidget {
-  final bool inBottomSheet;
-  final bool showInput;
-  final bool collapsibleModelSelector;
-
-  /// 隐藏尺寸设置区（官网式布局把尺寸置顶到面板外部渲染）
-  final bool showSizeSection;
-
-  /// 嵌入外部滚动容器：shrinkWrap + 禁用自身滚动 + 去掉内边距
-  final bool embedded;
-
-  /// 紧凑模式：缩小节间距
-  final bool compact;
-
-  const ParameterPanel({
-    super.key,
-    this.inBottomSheet = false,
-    this.showInput = false,
-    this.collapsibleModelSelector = false,
-    this.showSizeSection = true,
-    this.embedded = false,
-    this.compact = false,
-  });
+/// 参数面板组件（经典布局与移动端使用）
+///
+/// 由 generation_param_sections.dart 中的分节控件组合而成，
+/// 官网式布局的一体滚动列复用同一批分节控件。
+class ParameterPanel extends ConsumerWidget {
+  const ParameterPanel({super.key});
 
   @override
-  ConsumerState<ParameterPanel> createState() => _ParameterPanelState();
-}
-
-class _ParameterPanelState extends ConsumerState<ParameterPanel> {
-  late final TextEditingController _seedController;
-  late final FocusNode _seedFocusNode;
-
-  /// 节间距：紧凑模式下略微收紧
-  double get _sectionGap => widget.compact ? 10.0 : 16.0;
-
-  @override
-  void initState() {
-    super.initState();
-    _seedController = TextEditingController();
-    _seedFocusNode = FocusNode();
-  }
-
-  @override
-  void dispose() {
-    _seedController.dispose();
-    _seedFocusNode.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final params = ref.watch(
-      generationParamsNotifierProvider.select(selectParameterPanelViewData),
-    );
-    final generationState = ref.watch(imageGenerationNotifierProvider);
-    final kritaBridgeState = ref.watch(kritaBridgeNotifierProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
     final theme = Theme.of(context);
-    final isLauncherGenerating = generationState.isGenerating;
-    final isGenerating =
-        isLauncherGenerating || kritaBridgeState.isBridgeGenerating;
-    final canSkipCurrentBatch = isLauncherGenerating &&
-        generationState.currentImage > 0 &&
-        generationState.totalImages > generationState.currentImage;
-    final generationProgressText =
-        '${generationState.currentImage}/${generationState.totalImages}';
-
-    _syncSeedController(params.seed);
+    final advancedOptionsExpanded = ref.watch(
+      generationParamsNotifierProvider
+          .select((params) => params.advancedOptionsExpanded),
+    );
 
     return ListView(
-      padding: widget.embedded
-          ? EdgeInsets.zero
-          : EdgeInsets.all(widget.inBottomSheet ? 16 : 12),
-      shrinkWrap: widget.inBottomSheet || widget.embedded,
-      physics: widget.embedded
-          ? const NeverScrollableScrollPhysics()
-          : widget.inBottomSheet
-              ? const ClampingScrollPhysics()
-              : null,
+      padding: const EdgeInsets.all(12),
       children: [
-        // 提示词输入 (仅当 showInput 为 true 时显示)
-        if (widget.showInput) ...[
-          const PromptInputWidget(compact: false),
-          SizedBox(height: _sectionGap),
-
-          // 生成按钮
-          SizedBox(
-            height: 48,
-            child: isLauncherGenerating
-                ? Row(
-                    children: [
-                      if (canSkipCurrentBatch) ...[
-                        Expanded(
-                          child: ThemedButton(
-                            onPressed: () => ref
-                                .read(imageGenerationNotifierProvider.notifier)
-                                .skipCurrentRequest(),
-                            icon: const Icon(Icons.skip_next),
-                            label: Text(
-                              '${context.l10n.generation_skipCurrentBatch} $generationProgressText',
-                            ),
-                            style: ThemedButtonStyle.outlined,
-                          ),
-                        ),
-                        const SizedBox(width: 8),
-                      ],
-                      Expanded(
-                        child: ThemedButton(
-                          onPressed: () => ref
-                              .read(imageGenerationNotifierProvider.notifier)
-                              .cancel(),
-                          icon: const Icon(Icons.stop_circle_outlined),
-                          label: Text(
-                            context.l10n.generation_stopAllGeneration,
-                          ),
-                          style: ThemedButtonStyle.outlined,
-                        ),
-                      ),
-                    ],
-                  )
-                : ThemedButton(
-                    onPressed: isGenerating
-                        ? null
-                        : () async {
-                            if (ref
-                                .read(kritaBridgeNotifierProvider)
-                                .isBridgeGenerating) {
-                              AppToast.warning(
-                                context,
-                                context.l10n.kritaBridge_busyGenerating,
-                              );
-                              return;
-                            }
-                            if (params.prompt.isEmpty) {
-                              AppToast.info(
-                                context,
-                                context.l10n.generation_pleaseInputPrompt,
-                              );
-                              return;
-                            }
-                            final confirmed =
-                                await AssetProtectionGuard.confirmHighAnlasCost(
-                              context: context,
-                              ref: ref,
-                            );
-                            if (!confirmed || !context.mounted) {
-                              return;
-                            }
-                            ref
-                                .read(imageGenerationNotifierProvider.notifier)
-                                .generate(
-                                  ref.read(generationParamsNotifierProvider),
-                                );
-                          },
-                    icon: isGenerating ? null : const Icon(Icons.auto_awesome),
-                    isLoading: isGenerating,
-                    label: Text(
-                      isGenerating
-                          ? context.l10n.generation_generating
-                          : context.l10n.generation_generateImage,
-                    ),
-                    style: ThemedButtonStyle.filled,
-                  ),
-          ),          const SizedBox(height: 24),
-          const ThemedDivider(),
-          SizedBox(height: _sectionGap),
-        ],
-
         // 模型选择
-        if (widget.collapsibleModelSelector)
-          CollapsibleModelSelector(
-            title: context.l10n.generation_model,
-            currentModelName:
-                ImageModels.modelDisplayNames[params.model] ?? params.model,
-            initiallyExpanded:
-                ref.read(layoutStateNotifierProvider).webModelDrawerExpanded,
-            onExpansionChanged: (expanded) => ref
-                .read(layoutStateNotifierProvider.notifier)
-                .setWebModelDrawerExpanded(expanded),
-            child: _buildModelDropdown(),
-          )
-        else ...[
-          _buildSectionTitle(theme, context.l10n.generation_model),
-          const SizedBox(height: 8),
-          _buildModelDropdown(),
-        ],
+        const ModelSection(),
 
-        SizedBox(height: _sectionGap),
+        const SizedBox(height: 16),
 
-        // 尺寸设置（官网式布局中由外部置顶渲染）
-        if (widget.showSizeSection) ...[
-          _buildSectionTitle(theme, context.l10n.generation_imageSize),
-          const SizedBox(height: 8),
-          SizeSelector(
-            width: params.width,
-            height: params.height,
-            onChanged: (width, height) {
-              ref
-                  .read(generationParamsNotifierProvider.notifier)
-                  .updateSize(width, height);
-            },
-          ),
-          SizedBox(height: _sectionGap),
-        ],
+        // 尺寸设置
+        const SizeSection(),
+
+        const SizedBox(height: 16),
 
         // 采样器
-        _buildSectionTitle(theme, context.l10n.generation_sampler),
-        const SizedBox(height: 8),
-        ThemedDropdown<String>(
-          value: params.sampler,
-          items: Samplers.allSamplers.map((sampler) {
-            return DropdownMenuItem(
-              value: sampler,
-              child: Text(
-                Samplers.samplerDisplayNames[sampler] ?? sampler,
-                style: const TextStyle(fontSize: 13),
-              ),
-            );
-          }).toList(),
-          onChanged: (value) {
-            if (value != null) {
-              ref
-                  .read(generationParamsNotifierProvider.notifier)
-                  .updateSampler(value);
-            }
-          },
-        ),
+        const SamplerSection(),
 
-        SizedBox(height: _sectionGap),
+        const SizedBox(height: 16),
 
         // 调度器
-        _buildSectionTitle(theme, context.l10n.generation_noiseSchedule),
-        const SizedBox(height: 8),
-        ThemedDropdown<String>(
-          // V4/V4.5 模型不支持 native，如果当前值是 native 则显示 karras
-          value: params.isV4Model && params.noiseSchedule == 'native'
-              ? 'karras'
-              : params.noiseSchedule,
-          items: [
-            // V3 模型多一个 Native 选项
-            if (!params.isV4Model)
-              DropdownMenuItem(
-                value: 'native',
-                child: Text(
-                  NoiseSchedules.displayNames['native'] ?? 'Native',
-                  style: const TextStyle(fontSize: 13),
-                ),
-              ),
-            ...['karras', 'exponential', 'polyexponential'].map((schedule) {
-              return DropdownMenuItem(
-                value: schedule,
-                child: Text(
-                  NoiseSchedules.displayNames[schedule] ?? schedule,
-                  style: const TextStyle(fontSize: 13),
-                ),
-              );
-            }),
-          ],
-          onChanged: (value) {
-            if (value != null) {
-              ref
-                  .read(generationParamsNotifierProvider.notifier)
-                  .updateNoiseSchedule(value);
-            }
-          },
-        ),
+        const NoiseScheduleSection(),
 
-        SizedBox(height: _sectionGap),
+        const SizedBox(height: 16),
 
         // 步数
-        _buildSectionTitle(
-          theme,
-          context.l10n.generation_steps(params.steps.toString()),
-        ),
-        ThemedSlider(
-          value: params.steps.toDouble(),
-          min: 1,
-          max: 50,
-          divisions: 49,
-          onChanged: (value) {
-            ref
-                .read(generationParamsNotifierProvider.notifier)
-                .updateSteps(value.round());
-          },
-        ),
+        const StepsSection(),
 
         // CFG Scale
-        Row(
-          children: [
-            _buildSectionTitle(
-              theme,
-              context.l10n.generation_cfgScale(params.scale.toStringAsFixed(1)),
-            ),
-            const Spacer(),
-            // Decrisp (仅 V3 模型)
-            if (params.isV3Model) ...[
-              _ToggleButton(
-                label: 'Decrisp',
-                isEnabled: params.decrisp,
-                onChanged: (value) {
-                  ref
-                      .read(generationParamsNotifierProvider.notifier)
-                      .updateDecrisp(value);
-                },
-              ),
-              const SizedBox(width: 8),
-            ],
-            // Variety+ (所有模型)
-            _ToggleButton(
-              label: 'Variety+',
-              isEnabled: params.varietyPlus,
-              onChanged: (value) {
-                ref
-                    .read(generationParamsNotifierProvider.notifier)
-                    .updateVarietyPlus(value);
-              },
-            ),
-          ],
-        ),
-        ThemedSlider(
-          value: params.scale,
-          min: 1,
-          max: 20,
-          divisions: 190,
-          onChanged: (value) {
-            ref
-                .read(generationParamsNotifierProvider.notifier)
-                .updateScale(value);
-          },
-        ),
+        const CfgScaleSection(),
 
-        SizedBox(height: _sectionGap),
+        const SizedBox(height: 16),
 
         // 种子
-        _buildSectionTitle(theme, context.l10n.generation_seed),
-        const SizedBox(height: 8),
-        Row(
-          children: [
-            Expanded(
-              child: ThemedInput(
-                controller: _seedController,
-                focusNode: _seedFocusNode,
-                hintText: context.l10n.generation_seedRandom,
-                keyboardType: TextInputType.number,
-                onChanged: (value) {
-                  // 清空输入框时自动变成随机 (-1)
-                  final seed = value.isEmpty ? -1 : (int.tryParse(value) ?? -1);
-                  if (seed != params.seed) {
-                    ref
-                        .read(generationParamsNotifierProvider.notifier)
-                        .updateSeed(seed);
-                  }
-                },
-                suffixIcon: params.seed == -1
-                    ? null
-                    : Row(
-                        mainAxisSize: MainAxisSize.min,
-                        children: [
-                          // 复制按钮
-                          _SeedIconButton(
-                            icon: Icons.copy_rounded,
-                            tooltip: context.l10n.common_copy,
-                            onPressed: () {
-                              Clipboard.setData(
-                                ClipboardData(text: params.seed.toString()),
-                              );
-                              AppToast.success(
-                                context,
-                                context.l10n.common_copied,
-                              );
-                            },
-                          ),
-                          // 清空按钮
-                          _SeedIconButton(
-                            icon: Icons.clear_rounded,
-                            tooltip: context.l10n.common_clear,
-                            onPressed: () {
-                              _seedController.clear();
-                              ref
-                                  .read(
-                                    generationParamsNotifierProvider.notifier,
-                                  )
-                                  .updateSeed(-1);
-                            },
-                          ),
-                          const SizedBox(width: 4),
-                        ],
-                      ),
-              ),
-            ),
-            const SizedBox(width: 8),
-            // 种子锁定按钮
-            IconButton(
-              icon: Icon(
-                ref
-                        .watch(generationParamsNotifierProvider.notifier)
-                        .isSeedLocked
-                    ? Icons.lock
-                    : Icons.lock_open,
-                size: 20,
-              ),
-              onPressed: () {
-                // 先同步输入框的值到 state（防止用户输入后 state 未更新）
-                final inputText = _seedController.text.trim();
-                final inputSeed =
-                    inputText.isEmpty ? -1 : (int.tryParse(inputText) ?? -1);
-                if (inputSeed !=
-                    ref.read(generationParamsNotifierProvider).seed) {
-                  ref
-                      .read(generationParamsNotifierProvider.notifier)
-                      .updateSeed(inputSeed);
-                }
+        const SeedSection(),
 
-                ref
-                    .read(generationParamsNotifierProvider.notifier)
-                    .toggleSeedLock();
-                // 更新输入框显示
-                final newSeed = ref.read(generationParamsNotifierProvider).seed;
-                _seedController.text = newSeed == -1 ? '' : newSeed.toString();
-              },
-              tooltip: ref
-                      .watch(generationParamsNotifierProvider.notifier)
-                      .isSeedLocked
-                  ? context.l10n.generation_seedUnlock
-                  : context.l10n.generation_seedLock,
-              style: IconButton.styleFrom(
-                backgroundColor: ref
-                        .watch(generationParamsNotifierProvider.notifier)
-                        .isSeedLocked
-                    ? theme.colorScheme.primary.withValues(alpha: 0.15)
-                    : theme.colorScheme.surfaceContainerHighest,
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-              ),
-            ),
-          ],
-        ),
-
-        SizedBox(height: _sectionGap),
+        const SizedBox(height: 16),
 
         // ==================== 新功能面板 ====================
 
@@ -485,7 +80,7 @@ class _ParameterPanelState extends ConsumerState<ParameterPanel> {
         // Precise Reference 面板 (角色/风格参考)
         const PreciseReferencePanel(),
 
-        SizedBox(height: _sectionGap),
+        const SizedBox(height: 16),
 
         // 高级选项
         ExpansionTile(
@@ -494,463 +89,17 @@ class _ParameterPanelState extends ConsumerState<ParameterPanel> {
             style: theme.textTheme.titleSmall,
           ),
           tilePadding: EdgeInsets.zero,
-          initiallyExpanded: params.advancedOptionsExpanded,
+          initiallyExpanded: advancedOptionsExpanded,
           onExpansionChanged: (expanded) {
             ref
                 .read(generationParamsNotifierProvider.notifier)
                 .setAdvancedOptionsExpanded(expanded);
           },
-          children: [
-            // V3 模型: SMEA 选项 (非 DDIM 采样器时显示)
-            if (params.isV3Model && !params.sampler.contains('ddim')) ...[
-              // 标题和说明
-              Padding(
-                padding: const EdgeInsets.only(top: 4, bottom: 4),
-                child: Row(
-                  children: [
-                    Text(
-                      'SMEA',
-                      style: theme.textTheme.titleSmall?.copyWith(
-                        fontWeight: FontWeight.w600,
-                      ),
-                    ),
-                    const SizedBox(width: 8),
-                    Expanded(
-                      child: Text(
-                        '高分辨率采样优化',
-                        style: theme.textTheme.bodySmall?.copyWith(
-                          color: theme.colorScheme.onSurface
-                              .withValues(alpha: 0.6),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              // 选项行
-              Padding(
-                padding: const EdgeInsets.only(bottom: 4),
-                child: Row(
-                  children: [
-                    _SmeaAutoButton(
-                      isAuto: params.smeaAuto,
-                      onChanged: (value) {
-                        ref
-                            .read(generationParamsNotifierProvider.notifier)
-                            .updateSmeaAuto(value);
-                      },
-                    ),
-                    const SizedBox(width: 16),
-                    _SmeaOptions(
-                      smea: params.smea,
-                      smeaDyn: params.smeaDyn,
-                      isAutoEnabled: params.smeaAuto,
-                      onSmeaChanged: (value) {
-                        ref
-                            .read(generationParamsNotifierProvider.notifier)
-                            .updateSmea(value);
-                      },
-                      onSmeaDynChanged: (value) {
-                        ref
-                            .read(generationParamsNotifierProvider.notifier)
-                            .updateSmeaDyn(value);
-                      },
-                    ),
-                  ],
-                ),
-              ),
-              // Auto 模式说明 (仅 Auto 开启时显示)
-              if (params.smeaAuto)
-                Padding(
-                  padding: const EdgeInsets.only(bottom: 8),
-                  child: Text(
-                    context.l10n.generation_smeaDescription,
-                    style: theme.textTheme.bodySmall?.copyWith(
-                      color: theme.colorScheme.primary,
-                      fontStyle: FontStyle.italic,
-                    ),
-                  ),
-                ),
-            ],
-            // V4 模型: CFG Rescale
-            if (params.isV4Model)
-              ListTile(
-                contentPadding: EdgeInsets.zero,
-                title: Text(
-                  context.l10n.generation_cfgRescale(
-                    params.cfgRescale.toStringAsFixed(2),
-                  ),
-                ),
-                subtitle: ThemedSlider(
-                  value: params.cfgRescale,
-                  min: 0,
-                  max: 1,
-                  divisions: 100,
-                  onChanged: (value) {
-                    ref
-                        .read(generationParamsNotifierProvider.notifier)
-                        .updateCfgRescale(value);
-                  },
-                ),
-              ),
+          children: const [
+            AdvancedSamplingOptions(),
           ],
         ),
       ],
-    );
-  }
-
-  Widget _buildSectionTitle(ThemeData theme, String title) {
-    return Text(
-      title,
-      style: theme.textTheme.titleSmall?.copyWith(
-        fontWeight: FontWeight.w600,
-      ),
-    );
-  }
-
-  Widget _buildModelDropdown() {
-    final model = ref.watch(
-      generationParamsNotifierProvider.select((params) => params.model),
-    );
-    return ThemedDropdown<String>(
-      value: model,
-      items: ImageModels.allModels.map((model) {
-        return DropdownMenuItem(
-          value: model,
-          child: Text(
-            ImageModels.modelDisplayNames[model] ?? model,
-            style: const TextStyle(fontSize: 13),
-          ),
-        );
-      }).toList(),
-      onChanged: (value) {
-        if (value != null) {
-          ref
-              .read(generationParamsNotifierProvider.notifier)
-              .updateModel(value);
-        }
-      },
-    );
-  }
-
-  void _syncSeedController(int seed) {
-    final nextText = resolveSeedFieldSyncText(
-      currentText: _seedController.text,
-      seed: seed,
-      hasFocus: _seedFocusNode.hasFocus,
-    );
-    if (nextText == null) {
-      return;
-    }
-
-    _seedController.value = _seedController.value.copyWith(
-      text: nextText,
-      selection: TextSelection.collapsed(offset: nextText.length),
-      composing: TextRange.empty,
-    );
-  }
-}
-
-@visibleForTesting
-String? resolveSeedFieldSyncText({
-  required String currentText,
-  required int seed,
-  required bool hasFocus,
-}) {
-  if (hasFocus) {
-    return null;
-  }
-
-  final nextText = seed == -1 ? '' : seed.toString();
-  if (currentText == nextText) {
-    return null;
-  }
-
-  return nextText;
-}
-
-/// SMEA Auto 按钮 (V3 模型)
-class _SmeaAutoButton extends StatelessWidget {
-  final bool isAuto;
-  final ValueChanged<bool> onChanged;
-
-  const _SmeaAutoButton({
-    required this.isAuto,
-    required this.onChanged,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return Material(
-      color: isAuto
-          ? theme.colorScheme.primary.withValues(alpha: 0.15)
-          : theme.colorScheme.surfaceContainerHighest,
-      borderRadius: BorderRadius.circular(8),
-      child: InkWell(
-        onTap: () => onChanged(!isAuto),
-        borderRadius: BorderRadius.circular(8),
-        child: Container(
-          padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 10),
-          decoration: BoxDecoration(
-            border: Border.all(
-              color: isAuto
-                  ? theme.colorScheme.primary.withValues(alpha: 0.5)
-                  : theme.colorScheme.outline.withValues(alpha: 0.3),
-            ),
-            borderRadius: BorderRadius.circular(8),
-          ),
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Icon(
-                isAuto ? Icons.check_box : Icons.check_box_outline_blank,
-                size: 18,
-                color: isAuto
-                    ? theme.colorScheme.primary
-                    : theme.colorScheme.onSurface.withValues(alpha: 0.5),
-              ),
-              const SizedBox(width: 4),
-              Text(
-                'Auto',
-                style: TextStyle(
-                  fontSize: 13,
-                  fontWeight: FontWeight.w500,
-                  color: isAuto
-                      ? theme.colorScheme.primary
-                      : theme.colorScheme.onSurface.withValues(alpha: 0.7),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// SMEA 选项复选框 (V3 模型)
-class _SmeaOptions extends StatelessWidget {
-  final bool smea;
-  final bool smeaDyn;
-  final bool isAutoEnabled;
-  final ValueChanged<bool> onSmeaChanged;
-  final ValueChanged<bool> onSmeaDynChanged;
-
-  const _SmeaOptions({
-    required this.smea,
-    required this.smeaDyn,
-    required this.isAutoEnabled,
-    required this.onSmeaChanged,
-    required this.onSmeaDynChanged,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final isDisabled = isAutoEnabled;
-
-    return Row(
-      children: [
-        // SMEA 复选框
-        _buildCheckbox(
-          context: context,
-          label: 'SMEA',
-          value: smea,
-          isDisabled: isDisabled,
-          onChanged: onSmeaChanged,
-          theme: theme,
-        ),
-        const SizedBox(width: 16),
-        // DYN 复选框
-        _buildCheckbox(
-          context: context,
-          label: 'DYN',
-          value: smeaDyn,
-          isDisabled: isDisabled,
-          onChanged: onSmeaDynChanged,
-          theme: theme,
-        ),
-      ],
-    );
-  }
-
-  Widget _buildCheckbox({
-    required BuildContext context,
-    required String label,
-    required bool value,
-    required bool isDisabled,
-    required ValueChanged<bool> onChanged,
-    required ThemeData theme,
-  }) {
-    final color = isDisabled
-        ? theme.colorScheme.onSurface.withValues(alpha: 0.3)
-        : (value
-            ? theme.colorScheme.primary
-            : theme.colorScheme.onSurface.withValues(alpha: 0.7));
-
-    return InkWell(
-      onTap: isDisabled ? null : () => onChanged(!value),
-      borderRadius: BorderRadius.circular(4),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(
-            value ? Icons.check_box : Icons.check_box_outline_blank,
-            size: 20,
-            color: color,
-          ),
-          const SizedBox(width: 4),
-          Text(
-            label,
-            style: TextStyle(
-              fontSize: 13,
-              color: color,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/// 通用切换按钮 (用于 Variety+, Decrisp 等)
-class _ToggleButton extends StatefulWidget {
-  final String label;
-  final bool isEnabled;
-  final ValueChanged<bool> onChanged;
-
-  const _ToggleButton({
-    required this.label,
-    required this.isEnabled,
-    required this.onChanged,
-  });
-
-  @override
-  State<_ToggleButton> createState() => _ToggleButtonState();
-}
-
-class _ToggleButtonState extends State<_ToggleButton> {
-  bool _isHovered = false;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final isDark = theme.brightness == Brightness.dark;
-
-    // 计算背景色
-    Color backgroundColor;
-    if (widget.isEnabled) {
-      backgroundColor = _isHovered
-          ? theme.colorScheme.primary.withValues(alpha: 0.85)
-          : theme.colorScheme.primary;
-    } else {
-      final baseColor = isDark
-          ? Colors.white.withValues(alpha: 0.08)
-          : Colors.black.withValues(alpha: 0.05);
-      backgroundColor = _isHovered
-          ? (isDark
-              ? Colors.white.withValues(alpha: 0.15)
-              : Colors.black.withValues(alpha: 0.1))
-          : baseColor;
-    }
-
-    return MouseRegion(
-      onEnter: (_) => setState(() => _isHovered = true),
-      onExit: (_) => setState(() => _isHovered = false),
-      cursor: SystemMouseCursors.click,
-      child: GestureDetector(
-        onTap: () => widget.onChanged(!widget.isEnabled),
-        child: AnimatedContainer(
-          duration: const Duration(milliseconds: 150),
-          curve: Curves.easeInOut,
-          padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 5),
-          decoration: BoxDecoration(
-            color: backgroundColor,
-            borderRadius: BorderRadius.circular(16),
-            border: Border.all(
-              color: widget.isEnabled
-                  ? theme.colorScheme.primary
-                  : (_isHovered
-                      ? theme.colorScheme.outline.withValues(alpha: 0.4)
-                      : theme.colorScheme.outline.withValues(alpha: 0.2)),
-              width: 1,
-            ),
-          ),
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              if (widget.isEnabled) ...[
-                Icon(
-                  Icons.check_rounded,
-                  size: 14,
-                  color: theme.colorScheme.onPrimary,
-                ),
-                const SizedBox(width: 4),
-              ],
-              Text(
-                widget.label,
-                style: TextStyle(
-                  fontSize: 12,
-                  fontWeight: FontWeight.w500,
-                  color: widget.isEnabled
-                      ? theme.colorScheme.onPrimary
-                      : theme.colorScheme.onSurface
-                          .withValues(alpha: _isHovered ? 0.8 : 0.6),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// 种子输入框内的图标按钮
-class _SeedIconButton extends StatefulWidget {
-  final IconData icon;
-  final String tooltip;
-  final VoidCallback onPressed;
-
-  const _SeedIconButton({
-    required this.icon,
-    required this.tooltip,
-    required this.onPressed,
-  });
-
-  @override
-  State<_SeedIconButton> createState() => _SeedIconButtonState();
-}
-
-class _SeedIconButtonState extends State<_SeedIconButton> {
-  bool _isHovered = false;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return MouseRegion(
-      onEnter: (_) => setState(() => _isHovered = true),
-      onExit: (_) => setState(() => _isHovered = false),
-      cursor: SystemMouseCursors.click,
-      child: Tooltip(
-        message: widget.tooltip,
-        child: GestureDetector(
-          onTap: widget.onPressed,
-          child: Padding(
-            padding: const EdgeInsets.all(8),
-            child: Icon(
-              widget.icon,
-              size: 18,
-              color: _isHovered
-                  ? theme.colorScheme.primary
-                  : theme.colorScheme.onSurface.withValues(alpha: 0.5),
-            ),
-          ),
-        ),
-      ),
     );
   }
 }
