+++ lib/presentation/screens/generation/widgets/web_left_panel.dart
@@ -2,20 +2,26 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'package:nai_launcher/core/utils/localization_extension.dart';
+import '../../../../core/constants/api_constants.dart';
 import '../../../providers/image_generation_provider.dart';
 import '../../../providers/layout_state_provider.dart';
 import 'collapsed_panel.dart';
 import 'generation_controls/generation_controls.dart';
-import 'parameter_panel.dart';
+import 'generation_param_sections.dart';
+import 'img2img_panel.dart';
+import 'precise_reference_panel.dart';
 import 'prompt_input.dart';
-import 'size_selector.dart';
+import 'reverse_prompt_panel.dart';
+import 'unified_reference_panel.dart';
 
 /// 官网式布局左栏
 ///
-/// 一体式滚动列：尺寸设置置顶，提示词编辑器随内容自由增高（官网式），
-/// 角色卡片与其余参数依次排列在同一滚动域中；
-/// 生成控制条常驻钉底，不随内容滚动。
-class WebLeftPanel extends ConsumerWidget {
+/// 一体式滚动列，自上而下：尺寸 → 种子 → 提示词（随内容自由增高）→
+/// 反推/图生图/风格迁移/精准参考。
+/// 参数（模型/采样器/噪声调度/步数/CFG/CFG Rescale）是独立的二级菜单：
+/// 触发条固定在钉底生成控制条上方，点击后面板向上弹出、
+/// 悬浮于滚动内容之上（官网式，不与提示词等模块同层）。
+class WebLeftPanel extends ConsumerStatefulWidget {
   final ValueNotifier<bool> negativeModeNotifier;
   final bool isResizing;
 
@@ -26,7 +32,63 @@ class WebLeftPanel extends ConsumerWidget {
   });
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
+  ConsumerState<WebLeftPanel> createState() => _WebLeftPanelState();
+}
+
+class _WebLeftPanelState extends ConsumerState<WebLeftPanel>
+    with SingleTickerProviderStateMixin {
+  /// 参数二级菜单是否挂载（动画收起完毕后卸载，瞬态不持久化）
+  bool _paramsMenuMounted = false;
+
+  late final AnimationController _menuController = AnimationController(
+    vsync: this,
+    duration: const Duration(milliseconds: 220),
+  );
+  late final CurvedAnimation _menuAnimation = CurvedAnimation(
+    parent: _menuController,
+    curve: Curves.easeOutCubic,
+    reverseCurve: Curves.easeInCubic,
+  );
+  late final Animation<Offset> _menuSlide = Tween<Offset>(
+    begin: const Offset(0, 1),
+    end: Offset.zero,
+  ).animate(_menuAnimation);
+
+  @override
+  void initState() {
+    super.initState();
+    // 收起动画播完再卸载浮层
+    _menuController.addStatusListener((status) {
+      if (status == AnimationStatus.dismissed && mounted) {
+        setState(() => _paramsMenuMounted = false);
+      }
+    });
+  }
+
+  @override
+  void dispose() {
+    _menuAnimation.dispose();
+    _menuController.dispose();
+    super.dispose();
+  }
+
+  void _toggleParamsMenu() {
+    final opening = _menuController.status == AnimationStatus.dismissed ||
+        _menuController.status == AnimationStatus.reverse;
+    if (opening) {
+      setState(() => _paramsMenuMounted = true);
+      _menuController.forward();
+    } else {
+      _menuController.reverse();
+    }
+  }
+
+  void _closeParamsMenu() {
+    _menuController.reverse();
+  }
+
+  @override
+  Widget build(BuildContext context) {
     final theme = Theme.of(context);
     final layoutState = ref.watch(layoutStateNotifierProvider);
 
@@ -41,7 +103,7 @@ class WebLeftPanel extends ConsumerWidget {
     );
 
     final child = layoutState.webLeftPanelExpanded
-        ? _buildExpanded(context, ref, theme)
+        ? _buildExpanded(context, theme)
         : CollapsedPanel(
             icon: Icons.edit_note,
             label: context.l10n.generation_params,
@@ -51,7 +113,7 @@ class WebLeftPanel extends ConsumerWidget {
           );
 
     // 拖拽时不使用动画，避免粘滞感（与经典布局 LeftPanel 一致）
-    if (isResizing) {
+    if (widget.isResizing) {
       return Container(width: width, decoration: decoration, child: child);
     }
     return AnimatedContainer(
@@ -62,10 +124,10 @@ class WebLeftPanel extends ConsumerWidget {
     );
   }
 
-  Widget _buildExpanded(BuildContext context, WidgetRef ref, ThemeData theme) {
-    final size = ref.watch(
+  Widget _buildExpanded(BuildContext context, ThemeData theme) {
+    final modelName = ref.watch(
       generationParamsNotifierProvider.select(
-        (params) => (width: params.width, height: params.height),
+        (params) => ImageModels.modelDisplayNames[params.model] ?? params.model,
       ),
     );
 
@@ -99,49 +161,162 @@ class WebLeftPanel extends ConsumerWidget {
           ),
         ),
 
-        // 一体式滚动列
-        // 用 SingleChildScrollView 而非 ListView：列内容有限，且需要
-        // 保持子项 State（撤销历史、角色卡展开态）不随滚动销毁
+        // 内容区 + 参数二级菜单浮层
         Expanded(
-          child: SingleChildScrollView(
-            padding: const EdgeInsets.all(12),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: [
-                // 尺寸设置置顶：位于提示词上方，不会被长提示词推走
-                Text(
-                  context.l10n.generation_imageSize,
-                  style: theme.textTheme.titleSmall?.copyWith(
-                    fontWeight: FontWeight.w600,
+          child: LayoutBuilder(
+            builder: (context, constraints) {
+              return Stack(
+                children: [
+                  // 一体式滚动列
+                  // 用 SingleChildScrollView 而非 ListView：列内容有限，且需要
+                  // 保持子项 State（撤销历史、面板展开态）不随滚动销毁
+                  SingleChildScrollView(
+                    padding: const EdgeInsets.all(12),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.stretch,
+                      children: [
+                        // 尺寸设置置顶：不会被长提示词推走
+                        const SizeSection(),
+                        const SizedBox(height: 10),
+
+                        // 种子：位于尺寸与提示词之间
+                        const SeedSection(),
+                        const SizedBox(height: 10),
+
+                        // 提示词：随内容自由增高（官网式）
+                        // 多角色编辑入口在提示词工具栏的「角色」按钮
+                        PromptInputWidget(
+                          autoGrow: true,
+                          showMaximizeButton: false,
+                          negativeModeNotifier: widget.negativeModeNotifier,
+                        ),
+                        const SizedBox(height: 10),
+
+                        // 功能面板：反推 / 图生图 / 风格迁移 / 精准参考
+                        const ReversePromptPanel(),
+                        const SizedBox(height: 8),
+                        const Img2ImgPanel(),
+                        const SizedBox(height: 8),
+                        const UnifiedReferencePanel(),
+                        const SizedBox(height: 8),
+                        const PreciseReferencePanel(),
+                      ],
+                    ),
                   ),
-                ),
-                const SizedBox(height: 6),
-                SizeSelector(
-                  width: size.width,
-                  height: size.height,
-                  onChanged: (width, height) => ref
-                      .read(generationParamsNotifierProvider.notifier)
-                      .updateSize(width, height),
-                ),
-                const SizedBox(height: 10),
 
-                // 提示词：随内容自由增高（官网式）
-                // 多角色编辑入口在提示词工具栏的「角色」按钮，不再内嵌卡片
-                PromptInputWidget(
-                  autoGrow: true,
-                  showMaximizeButton: false,
-                  negativeModeNotifier: negativeModeNotifier,
-                ),
-                const SizedBox(height: 8),
+                  // 参数二级菜单：遮罩淡入淡出 + 抽屉从触发条上方拉出
+                  if (_paramsMenuMounted) ...[
+                    Positioned.fill(
+                      child: FadeTransition(
+                        opacity: _menuAnimation,
+                        child: GestureDetector(
+                          behavior: HitTestBehavior.opaque,
+                          onTap: _closeParamsMenu,
+                          child: ColoredBox(
+                            color: Colors.black.withValues(alpha: 0.25),
+                          ),
+                        ),
+                      ),
+                    ),
+                    Positioned(
+                      left: 0,
+                      right: 0,
+                      bottom: 0,
+                      child: ClipRect(
+                        child: SlideTransition(
+                          position: _menuSlide,
+                          child: ConstrainedBox(
+                            constraints: BoxConstraints(
+                              maxHeight: constraints.maxHeight - 12,
+                            ),
+                            child: Material(
+                              elevation: 8,
+                              borderRadius: const BorderRadius.vertical(
+                                top: Radius.circular(10),
+                              ),
+                              clipBehavior: Clip.antiAlias,
+                              color: theme.colorScheme.surface,
+                              child: Container(
+                                decoration: BoxDecoration(
+                                  borderRadius: const BorderRadius.vertical(
+                                    top: Radius.circular(10),
+                                  ),
+                                  border: Border(
+                                    top: BorderSide(color: theme.dividerColor),
+                                    left: BorderSide(color: theme.dividerColor),
+                                    right:
+                                        BorderSide(color: theme.dividerColor),
+                                  ),
+                                ),
+                                child: const SingleChildScrollView(
+                                  padding: EdgeInsets.all(12),
+                                  child: Column(
+                                    crossAxisAlignment:
+                                        CrossAxisAlignment.stretch,
+                                    children: [
+                                      ModelSection(),
+                                      SizedBox(height: 10),
+                                      SamplerSection(),
+                                      SizedBox(height: 10),
+                                      NoiseScheduleSection(),
+                                      SizedBox(height: 10),
+                                      StepsSection(),
+                                      CfgScaleSection(),
+                                      AdvancedSamplingOptions(),
+                                    ],
+                                  ),
+                                ),
+                              ),
+                            ),
+                          ),
+                        ),
+                      ),
+                    ),
+                  ],
+                ],
+              );
+            },
+          ),
+        ),
 
-                // 其余参数（尺寸已置顶不再重复；紧凑嵌入，模型收进抽屉）
-                const ParameterPanel(
-                  embedded: true,
-                  compact: true,
-                  showSizeSection: false,
-                  collapsibleModelSelector: true,
+        // 参数二级菜单触发条：固定在钉底控制条上方，永远一步可达
+        Material(
+          color: theme.colorScheme.surface,
+          child: InkWell(
+            onTap: _toggleParamsMenu,
+            child: Container(
+              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 10),
+              decoration: BoxDecoration(
+                border: Border(
+                  top: BorderSide(color: theme.dividerColor, width: 1),
                 ),
-              ],
+              ),
+              child: Row(
+                children: [
+                  Text(
+                    context.l10n.generation_params,
+                    style: theme.textTheme.titleSmall,
+                  ),
+                  const SizedBox(width: 8),
+                  Expanded(
+                    child: Text(
+                      modelName,
+                      style: theme.textTheme.bodySmall?.copyWith(
+                        color: theme.colorScheme.primary,
+                      ),
+                      overflow: TextOverflow.ellipsis,
+                      textAlign: TextAlign.end,
+                    ),
+                  ),
+                  const SizedBox(width: 4),
+                  // 箭头随抽屉开合旋转（收起朝上提示拉出方向）
+                  RotationTransition(
+                    turns: Tween<double>(begin: 0, end: 0.5)
+                        .animate(_menuAnimation),
+                    child: const Icon(Icons.keyboard_arrow_up, size: 20),
+                  ),
+                ],
+              ),
             ),
           ),
         ),
